* commit '354a75b2c6117583f88d63141772a959a4735a56':
  junit 5 dependency issue resolved so tests are running again
  plugin setting changed
  possible fix for coverage coming as zero not sure
  possible fix for coverage coming as zero not sure
  sonar issues fixed
  sonar issues fixed
  unused imports removed
  removed cloud.aws.region.static property setting
  lombok removal  related changes
  lombok removal  related changes
  wgu boot upgrade 1.3.4 and breaking changes fixes for nimbusds 9.4

+++ src/main/java/edu/wgu/dm/service/UserManagementService.java
@@ -1,10 +1,7 @@
 package edu.wgu.dm.service;
 
 import edu.wgu.boot.core.exception.AuthorizationException;
-import edu.wgu.dm.dto.response.BulkCreateResponse;
-import edu.wgu.dm.dto.security.BulkUsers;
 import edu.wgu.dm.dto.security.Person;
-import edu.wgu.dm.dto.security.Role;
 import edu.wgu.dm.dto.security.User;
 import edu.wgu.dm.dto.security.UserSummary;
 import edu.wgu.dm.exception.UserNotFoundException;
@@ -12,37 +9,42 @@ import edu.wgu.dm.repository.RoleRepo;
 import edu.wgu.dm.repository.UserRepo;
 import edu.wgu.dm.service.feign.PersonService;
 import edu.wgu.dm.util.Permissions;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
-import lombok.AccessLevel;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 
 @Service
-@Slf4j
-@RequiredArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class UserManagementService {
 
     private final UserRepo adminRepo;
-
     private final RoleRepo roleRepo;
+    private final PersonService personService;
+    private static final String NON_NULL_USER_ID_IS_REQUIRED = "non null userId is required";
 
-    PersonService personService;
+    public UserManagementService(UserRepo adminRepo, RoleRepo roleRepo,
+        PersonService personService) {
+        this.adminRepo = adminRepo;
+        this.roleRepo = roleRepo;
+        this.personService = personService;
+    }
 
-    public User getUser(@NonNull String userId) {
+    public User getUser(String userId) {
+        if (userId == null) {
+            throw new NullPointerException(NON_NULL_USER_ID_IS_REQUIRED);
+        }
         return this.adminRepo.getUserById(userId)
                              .orElseThrow(() -> new UserNotFoundException(userId));
     }
 
-    public void saveUser(@NonNull String userId, @NonNull User user) {
+    public void saveUser(String userId, User user) {
+        if (userId == null) {
+            throw new NullPointerException(NON_NULL_USER_ID_IS_REQUIRED);
+        }
+        if (user == null) {
+            throw new NullPointerException("non null user is required");
+        }
         Set<Long> roles = user.getRoleIds();
 
         checkIfSystemUser(roles, userId);
@@ -50,9 +52,12 @@ public class UserManagementService {
     }
 
     /**
-     *  We will need to user deletion as a soft delete
-      */
-    public void deleteUser(@NonNull String userId) {
+     * We will need to user deletion as a soft delete
+     */
+    public void deleteUser(String userId) {
+        if (userId == null) {
+            throw new NullPointerException(NON_NULL_USER_ID_IS_REQUIRED);
+        }
         this.adminRepo.deleteUser(userId);
     }
 
@@ -60,14 +65,21 @@ public class UserManagementService {
         return this.adminRepo.getAllUsers();
     }
 
-    public List<UserSummary> getUsersForTask(@NonNull Long taskId) {
+    public List<UserSummary> getUsersForTask(Long taskId) {
+        if (taskId == null) {
+            throw new NullPointerException("non null TaskId is required");
+        }
         return this.adminRepo.getUsersByTask(taskId);
     }
 
-    public User createUser(@NonNull String userId) {
+    public User createUser(String userId) {
+        if (userId == null) {
+            throw new NullPointerException(NON_NULL_USER_ID_IS_REQUIRED);
+        }
         Person person = this.personService.getPersonByUsername(userId);
-        if (Boolean.FALSE.equals(person.getIsEmployee()))
+        if (Boolean.FALSE.equals(person.getIsEmployee())) {
             throw new IllegalArgumentException("User is not an employee.");
+        }
 
         return this.adminRepo.getUserById(person.getUserId())
                              .orElseGet(() -> saveUser(person));
@@ -75,57 +87,32 @@ public class UserManagementService {
 
     private User saveUser(Person person) {
         User user = new User(person);
-        return this.adminRepo.saveUser(user).orElseThrow(()->new IllegalStateException("Saving User failed"));
+        return this.adminRepo.saveUser(user)
+                             .orElseThrow(() -> new IllegalStateException("Saving User failed"));
     }
 
-    public BulkCreateResponse createUsers(@NonNull String userId, @NonNull BulkUsers users) {
-        List<User> toCreate = new ArrayList<>();
-        List<String> failed = new ArrayList<>();
-
-        List<Role> roles = users.getRoles()
-                                .stream()
-                                .map(Role::new)
-                                .collect(Collectors.toList());
-
-        checkIfSystemUser(users.getRoles(), userId);
-        users.getUsernames()
-             .forEach(name -> {
-                 try {
-                     Person person = this.personService.getPersonByUsername(name);
-                     User user = this.adminRepo.getUserById(person.getUserId())
-                                               .orElseGet(() -> new User(person));
-                     user.getRoles()
-                         .addAll(roles);
-                     user.getTasks()
-                         .addAll(users.getTasks());
-                     toCreate.add(user);
-                 } catch (Exception e) {
-                     log.error("Error creating new user.", e);
-                     failed.add(name);
-                 }
-             });
-
-        this.adminRepo.saveUsers(toCreate);
-        return new BulkCreateResponse(toCreate, failed);
-    }
 
     /**
-     * Validate that only a system user can assign system role to other user. If any of the incoming
-     * role IDs match with any role that has system permission, then check to see if the current user
-     * has the SYSTEM permission. If not, throw an AuthorizationException.
-     *  @param roleIds
-     * @param userId
-     * @return
+     * Validate that only a system user can assign system role to other user. If any of the incoming role IDs match with
+     * any role that has system permission, then check to see if the current user has the SYSTEM permission. If not,
+     * throw an AuthorizationException.
+     *
+     * @param roleIds List of role ids
+     * @param userId User id to check
      */
-    private UserSummary checkIfSystemUser(@NonNull Collection<Long> roleIds, @NonNull String userId) {
-        List<Long> rolesWithSystem = this.roleRepo.getRolesByPermission(Permissions.SYSTEM);
-        UserSummary summary = null;
-        if (CollectionUtils.containsAny(roleIds, rolesWithSystem)) {
-            summary= this.adminRepo.getUserWithPermission(userId, Permissions.SYSTEM)
-                          .orElseThrow(
-                                  () -> new AuthorizationException("Only SYSTEM users can assign SYSTEM permissions"));
+    private void checkIfSystemUser(Collection<Long> roleIds, String userId) {
+        if (roleIds == null) {
+            throw new NullPointerException("non null roleIds is required");
         }
-
-        return summary;
+        if (userId == null) {
+            throw new NullPointerException(NON_NULL_USER_ID_IS_REQUIRED);
+        }
+        List<Long> rolesWithSystem = this.roleRepo.getRolesByPermission(Permissions.SYSTEM);
+        if (CollectionUtils.containsAny(roleIds, rolesWithSystem)) {
+            int count = this.adminRepo.checkIfUserHasPermission(userId, Permissions.SYSTEM);
+            if (count == 0) {
+                throw new AuthorizationException("Only SYSTEM users can assign SYSTEM permissions");
+            }
+        }
     }
 }
